geolocation-API/getCurrentPosition_permission_allow.https.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Geolocation Test: getCurrentPosition() location access</title>
<link
  rel="help"
  href="https://www.w3.org/TR/geolocation/#getcurrentposition-method"
/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
  promise_test(async (t) => {
    t.add_cleanup(() => {
      return test_driver.set_permission({ name: "geolocation" }, "prompt");
    });
    await test_driver.set_permission({ name: "geolocation" }, "granted");
    const promise = new Promise((resolve, reject) => {
      navigator.geolocation.getCurrentPosition(resolve, reject);
    });
    const result = await promise;
    assert_true(
      result instanceof GeolocationPosition,
      "Expected GeolocationPosition"
    );
  }, "User allows access, check that success callback is called.");
</script>
